pyproject.toml
[project]
name = "chuk-mcp-music"
version = "0.1.0"
description = "Music design system for MCP - shadcn/ui for music composition"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Chris Hay", email = "chrishayuk@somejunkmailbox.com" }
]
keywords = [
    "mcp",
    "music",
    "midi",
    "composition",
    "design-system",
    "llm",
    "ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mido>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "chuk-mcp-server>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
chuk-music = "chuk_mcp_music.cli:main"

[project.entry-points."chuk.mcp.servers"]
music = "chuk_mcp_music.server:mcp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chuk_mcp_music"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
